+++ CHANGELOG.md
@@ -145,6 +145,7 @@
 
 ## 3.0.0
 * Added option to hide observation and forecast values 
+* Added option to have more detailed condition categories
 * Changed readme to improve comprehensibility
 * Changed naming of some parameters (all backwards compatible, no need to update your config)
 * Fixed unsupported forecast day would throw an error in the darksky api